requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
dnspython==2.8.0
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.8.1
pydantic==2.12.4
pydantic_core==2.41.5
pymongo==4.15.4
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
sniffio==1.3.1
tavily-python==0.7.13
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
